Conflicts:
	site/frontend/widgets/socialLike/views/_yh_min.php
	site/frontend/widgets/socialLike/views/simple.php
	site/frontend/www-submodule

+++ site/console/commands/UsersCommand.php
@@ -46,41 +46,41 @@ class UsersCommand extends CConsoleCommand
 
         $criteria = new CDbCriteria;
         $criteria->condition = '`group` = 0 AND last_active < :last_active AND login_date < :last_active
-         AND (register_date IS NULL OR register_date < :last_active)';
+         AND (register_date IS NULL OR register_date < :last_active) AND id < 14000';
         $criteria->params = array(':last_active' => date("Y-m-d H:i:s", strtotime('-2 month')));
+        $criteria->limit = 1000;
 
         $users = User::model()->findAll($criteria);
         $k = 0;
         foreach ($users as $user) {
             if ($user->communityContentsCount == 0 && $user->cookRecipesCount == 0
-                && $user->recipeBookRecipesCount == 0
-            ){
-                Yii::app()->db->createCommand()->delete('users', 'id='.$user->id);
-//                echo $user->id."\n";
-            $k++;
-    }
+                && $user->recipeBookRecipesCount == 0 && $user->commentsCount == 0
+                && $user->photosCount == 0
+            ) {
+                Yii::app()->db->createCommand()->delete('users', 'id=' . $user->id);
+                $k++;
+            }
         }
 
-        echo $k."\n";
+        echo $k . "\n";
 
 
         $criteria = new CDbCriteria;
-        $criteria->condition = '`group` = 0 AND last_active IS NULL AND login_date = "0000-00-00 00:00:00"
-        ';
+        $criteria->condition = '`group` = 0 AND last_active IS NULL AND login_date = "0000-00-00 00:00:00" AND id < 14000';
         //$criteria->params = array(':last_active' => date("Y-m-d H:i:s", strtotime('-2 month')));
 
         $users = User::model()->findAll($criteria);
         $k = 0;
         foreach ($users as $user) {
             if ($user->communityContentsCount == 0 && $user->cookRecipesCount == 0
-                && $user->recipeBookRecipesCount == 0
-            ){
-                Yii::app()->db->createCommand()->delete('users', 'id='.$user->id);
-//                echo $user->id."\n";
-            $k++;
+                && $user->recipeBookRecipesCount == 0 && $user->commentsCount == 0
+                && $user->photosCount == 0
+            ) {
+                Yii::app()->db->createCommand()->delete('users', 'id=' . $user->id);
+                $k++;
             }
         }
 
-        echo $k."\n";
+        echo $k . "\n";
     }
 }
